Separated array header from array data, see #97
This change now allows implementations of Array#push and similar as the array reference remains the same. Nice side effect is that Array#slice can return a view on the original array's data as well, though this requires meticulously keeping track of what's a slice and what's not when disposing, for now (will get better with GC).